+++ src/Data/Tuple/Constraint/TypeSynonym.hs
@@ -1,14 +1,12 @@
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE KindSignatures #-}
-#if __GLASGOW_HASKELL__ >= 800
+{-# LANGUAGE Safe #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE PolyKinds #-}
-# if __GLASGOW_HASKELL__ < 806
+#if __GLASGOW_HASKELL__ < 806
 {-# LANGUAGE TypeInType #-}
-# endif
 #endif
-{-# LANGUAGE Safe #-}
 
 -- | This module provides type aliases that emulate the behavior of GHC's constraint
 -- tuple syntax. Unlike GHC's built-in constraint tuples, the type aliases in this
@@ -16,12 +14,7 @@
 --
 -- The type aliases in this module are defined by way of type synonyms that
 -- decompose applications of constraint tuple type constructors to their arguments.
--- This requires the use of GHC capabilities that are only present on GHC 8.0 or
--- later, so this module does not export anything on earlier versions of GHC.
 module Data.Tuple.Constraint.TypeSynonym
-#if __GLASGOW_HASKELL__ < 800
-  () where
-#else
   ( -- * Constraint tuples
     CTuple0
   , CTuple1
@@ -358,4 +351,3 @@ type CTuple64 = Decomposer64 ('Proxy :: Proxy (((), (), (), (), (), (), (), (), 
 type Decomposer64 (x :: Proxy ((f :: Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint -> Constraint) c1 c2 c3 c4 c5 c6 c7 c8 c9 c10 c11 c12 c13 c14 c15 c16 c17 c18 c19 c20 c21 c22 c23 c24 c25 c26 c27 c28 c29 c30 c31 c32 c33 c34 c35 c36 c37 c38 c39 c40 c41 c42 c43 c44 c45 c46 c47 c48 c49 c50 c51 c52 c53 c54 c55 c56 c57 c58 c59 c60 c61 c62 c63 c64)) = f
 #endif
 
-#endif